chore: rename parser package to evaluator
We keep the parser package for our generated lexer, parser and base
classes.
This shows that the visitor/listener are the step after parsing.
