+++ src/main/java/nl/arothuis/antlr4calculator/core/evaluator/CalculationListener.java
@@ -1,6 +1,8 @@
-package nl.arothuis.antlr4calculator.core.parser;
+package nl.arothuis.antlr4calculator.core.evaluator;
 
 import java.util.Stack;
+import nl.arothuis.antlr4calculator.core.parser.CalculatorBaseListener;
+import nl.arothuis.antlr4calculator.core.parser.CalculatorParser;
 
 public class CalculationListener extends CalculatorBaseListener {
     /**
